requirements.txt
# AI SecOps Tool Dependencies

# Microsoft PyRIT
pyrit>=0.1.0
rpyc>=5.0.0  # Required for PyRIT Gradio GUI
gradio>=4.0.0  # Required for PyRIT Gradio GUI

# IBM Adversarial Robustness Toolbox
adversarial-robustness-toolbox>=1.18.0

# Google Cloud / Vertex AI
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0

# General utilities
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0

# Testing and analysis
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0

